User domain

+++ src/routers/Commons.kt
@@ -3,4 +3,4 @@ package com.example.routers
 import io.ktor.locations.Location
 
 @Location("{id}")
-data class IntId(val id: Int)
+data class IntId(val id: Long)
